+++ CHANGES.rst
@@ -34,6 +34,7 @@ Release History
 **Changed**
 
 - Increased minimum TensorFlow version to 1.2.0
+- Improved error checking for input/target data
 
 **Fixed**